[1.1.0] Updated specs
