+++ spec/arfi/sql_function_loader_spec.rb
@@ -3,14 +3,6 @@
 RSpec.describe Arfi::SqlFunctionLoader, :pgsql do
   include ArfiSpec::TmpRoot
 
-  def load!
-    described_class.load!(verbose: false)
-  end
-
-  def select_value(sql)
-    ActiveRecord::Base.connection.select_value(sql)
-  end
-
   before do
     write_function('db/functions/public/arfi_echo.sql', <<~SQL)
       CREATE OR REPLACE FUNCTION arfi_echo() RETURNS text
@@ -19,7 +11,7 @@ RSpec.describe Arfi::SqlFunctionLoader, :pgsql do
   end
 
   it 'loads generic functions from db/functions' do
-    load!
+    load_functions!
     expect(select_value('SELECT arfi_echo()')).to eq('generic')
   end
 
@@ -32,7 +24,7 @@ RSpec.describe Arfi::SqlFunctionLoader, :pgsql do
     end
 
     it 'prefers adapter-specific override db/functions/postgresql over db/functions' do
-      load!
+      load_functions!
       expect(select_value('SELECT arfi_echo()')).to eq('postgres')
     end
   end
@@ -47,7 +39,7 @@ RSpec.describe Arfi::SqlFunctionLoader, :pgsql do
     end
 
     it 'ignores underscore-prefixed sql files' do
-      load!
+      load_functions!
       expect(select_value('SELECT arfi_ok()')).to eq(1)
     end
   end
