+++ spec/arfi/sql_trigger_loader_mysql_spec.rb
@@ -3,14 +3,6 @@
 RSpec.describe Arfi::SqlTriggerLoader, :mysql do
   include ArfiSpec::TmpRoot
 
-  def load!
-    described_class.load!(verbose: false)
-  end
-
-  def select_value(sql)
-    ActiveRecord::Base.connection.select_value(sql)
-  end
-
   before do
     write_function('db/functions/public/users.sql', <<~SQL)
       CREATE TABLE IF NOT EXISTS users (id serial PRIMARY KEY, name varchar(255));
@@ -18,25 +10,16 @@ RSpec.describe Arfi::SqlTriggerLoader, :mysql do
     Arfi::SqlFunctionLoader.load!(verbose: false)
   end
 
-  def mysql_trigger_sql(name)
-    <<~SQL
-      CREATE TRIGGER #{name}
-        BEFORE INSERT ON users
-        FOR EACH ROW
-        SET NEW.name = 'triggered';
-    SQL
-  end
-
   it 'loads mysql triggers from db/triggers/mysql/public' do
     write_function('db/triggers/mysql/public/arfi_before_ins.sql', mysql_trigger_sql('arfi_before_ins'))
-    load!
+    load_triggers!
     expect(ActiveRecord::Base.trigger_exists?('users', 'arfi_before_ins')).to be(true)
   end
 
   it 'ignores underscore-prefixed files' do
     write_function('db/triggers/mysql/public/_boom.sql', 'SELECT 1/0;')
     write_function('db/triggers/mysql/public/arfi_before_ins.sql', mysql_trigger_sql('arfi_before_ins'))
-    load!
+    load_triggers!
     expect(ActiveRecord::Base.trigger_exists?('users', 'arfi_before_ins')).to be(true)
   end
 
@@ -57,7 +40,7 @@ RSpec.describe Arfi::SqlTriggerLoader, :mysql do
     end
 
     it 'prefers adapter file over generic' do
-      load!
+      load_triggers!
       expect(ActiveRecord::Base.trigger_exists?('users', 'arfi_echo')).to be(true)
     end
   end
